+++ pkg/plan/diff.go
@@ -10,6 +10,7 @@ import (
 
 func (p *Plan) Diff(b *Plan, diffWide int) {
 	var visited []string
+	noChanges := 0
 
 	for _, rel := range p.body.Releases {
 		m := rel.Uniq() + ".yml"
@@ -20,15 +21,22 @@ func (p *Plan) Diff(b *Plan, diffWide int) {
 
 		change := diff.Manifests(oldSpecs, newSpecs, []string{}, true, diffWide, os.Stdout)
 		if !change {
+			noChanges++
 			log.Info("❎  ", rel.Uniq(), " no changes")
 		}
 	}
 
+	previous := false
 	for _, rel := range b.body.Releases {
 		m := string(rel.Uniq() + ".yml")
 		if !helper.Contains(m, visited) {
+			previous = true
 			log.Warn(rel.Uniq(), " was found in previous planfile but not affected in new")
 		}
+	}
+
+	if noChanges == len(p.body.Releases) && !previous {
+		log.Info("Plan has no changes")
 	}
 }
