+++ pkg/template/func.go
@@ -1,13 +1,9 @@
 package template
 
 import (
-	"bytes"
-	"io/ioutil"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
-	"github.com/helmwave/helmwave/pkg/helper"
-	log "github.com/sirupsen/logrus"
 )
 
 var (
@@ -28,44 +24,6 @@ var (
 		"hasKey":         HasKey,
 	}
 )
-
-func Tpl2yml(tpl, yml string, data interface{}) error {
-	log.WithFields(log.Fields{
-		"from": tpl,
-		"to":   yml,
-	}).Debug("📄 Render file")
-
-	if data == nil {
-		data = map[string]interface{}{}
-	}
-
-	src, err := ioutil.ReadFile(tpl)
-	if err != nil {
-		return err
-	}
-
-	// Template
-	t := template.Must(template.New("tpl").Funcs(FuncMap()).Parse(string(src)))
-	var buf bytes.Buffer
-	err = t.Execute(&buf, data)
-	if err != nil {
-		return err
-	}
-
-	log.Debug(yml, " contents\n", buf.String())
-
-	f, err := helper.CreateFile(yml)
-	if err != nil {
-		return err
-	}
-
-	_, err = f.WriteString(buf.String())
-	if err != nil {
-		return err
-	}
-
-	return f.Close()
-}
 
 func FuncMap() template.FuncMap {
 	funcMap := sprig.TxtFuncMap()
